casa_de_mi_padre/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
firebase-admin==6.3.0
Flask==3.0.0
Flask-RESTful==0.3.10
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
lxml==5.1.0
MarkupSafe==2.1.3
msgpack==1.0.7
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
typing_extensions==4.9.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2
